pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neckenml-analyzer"
version = "0.1.0"
description = "Swedish folk music audio analysis and dance style classification"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "NeckenML Contributors"}
]
keywords = ["audio", "music", "folk", "swedish", "classification", "dance", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "joblib>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.1.0",
]

[project.optional-dependencies]
audio = [
    "essentia-tensorflow>=2.1b6.dev0",
    "librosa>=0.9.0",
    "madmom>=0.16.0",
    "soundfile>=0.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/svnoak/neckenml-analyzer"
Documentation = "https://github.com/yourusername/neckenml-analyzer/blob/main/README.md"
Repository = "https://github.com/yourusername/neckenml-analyzer"
Issues = "https://github.com/yourusername/neckenml-analyzer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["neckenml*"]

[tool.setuptools.package-data]
neckenml = ["models/pretrained/*.pkl"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
